requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pdfplumber>=0.10.3
python-dotenv>=1.0.0
pinecone-client>=3.0.0
langchain>=0.1.0
langchain-community>=0.0.10
openai>=1.12.0
langchain-openai>=0.0.5
tenacity>=8.2.3 